+++ CHANGELOG.md
@@ -6,6 +6,13 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 None
+<a name="4.0.2"></a>
+
+## [4.0.2] 2022-11-01
+### Added
+- Add `offnet` property in IncomingNumber resource
+- Add `offnet` query parameter for listIncomingNumbers
+
 <a name="4.0.1"></a>
 ## [4.0.1] 2022-08-19
 ### Added